requirements.txt
click==8.1.3
et-xmlfile==1.1.0
mypy-extensions==0.4.3
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.1
pathspec==0.10.2
platformdirs==2.5.4
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
tomli==2.0.1
